refactor index file for easier testing. refactor routes for organization
