+++ backend/src/routes/game.ts
@@ -1,9 +1,9 @@
 import express from 'express';
 import auth from '../utils/auth.js';
-import placeBet from './game-place-bet.js';
-import getWinStreaks from './game-win-streaks.js';
-import getRecentRolls from './game-recent-rolls.js';
-import getCurrentGame from './game-current.js';
+import placeBet from '../api/game-place-bet.js';
+import getWinStreaks from '../api/game-win-streaks.js';
+import getRecentRolls from '../api/game-recent-rolls.js';
+import getCurrentGame from '../api/game-current.js';
 
 const router = express.Router();
